+++ integration-test/Analysis/MavenSpec.hs
@@ -3,12 +3,31 @@
 
 module Analysis.MavenSpec (spec) where
 
-import Analysis.FixtureExpectationUtils
-import Analysis.FixtureUtils
-import Path
+import Analysis.FixtureExpectationUtils (
+  DependencyResultsSummary (
+    DependencyResultsSummary,
+    graphType,
+    numDeps,
+    numDirectDeps,
+    numEdges,
+    numManifestFiles
+  ),
+  expectProject,
+  testSuiteDepResultSummary,
+  withAnalysisOf,
+ )
+import Analysis.FixtureUtils (
+  AnalysisTestFixture (AnalysisTestFixture),
+  FixtureArtifact (FixtureArtifact),
+  FixtureEnvironment (NixEnv),
+ )
+import Path (reldir)
 import Strategy.Maven qualified as Maven
-import Test.Hspec
-import Types
+import Test.Hspec (Spec, aroundAll, describe, it)
+import Types (
+  DiscoveredProjectType (MavenProjectType),
+  GraphBreadth (Complete),
+ )
 
 mavenEnv :: FixtureEnvironment
 mavenEnv = NixEnv ["maven", "jdk11"]
@@ -25,13 +44,39 @@ keycloak =
       [reldir|maven/keycloak/|]
       [reldir|keycloak-15.1.0//|]
 
+guava :: AnalysisTestFixture Maven.MavenProject
+guava =
+  AnalysisTestFixture
+    "guava"
+    Maven.discover
+    mavenEnv
+    Nothing
+    $ FixtureArtifact
+      "https://github.com/google/guava/archive/refs/tags/v31.1.tar.gz"
+      [reldir|maven/guava/|]
+      [reldir|guava-31.1|]
+
 testKeycloak :: Spec
-testKeycloak =
+testKeycloak = do
   aroundAll (withAnalysisOf keycloak) $ do
     describe "keycloak" $ do
       it "should find targets" $ \(result, extractedDir) ->
         expectProject (MavenProjectType, extractedDir) result
 
+testGuava :: Spec
+testGuava =
+  testSuiteDepResultSummary
+    guava
+    Types.MavenProjectType
+    DependencyResultsSummary
+      { numDeps = 85
+      , numDirectDeps = 17
+      , numEdges = 75
+      , numManifestFiles = 1
+      , graphType = Complete
+      }
+
 spec :: Spec
 spec = do
   testKeycloak
+  testGuava