+++ integration-test/Analysis/RustSpec.hs
@@ -3,12 +3,18 @@
 
 module Analysis.RustSpec (spec) where
 
-import Analysis.FixtureExpectationUtils
-import Analysis.FixtureUtils
-import Path
+import Analysis.FixtureExpectationUtils (
+  testSuiteHasSomeDepResults,
+ )
+import Analysis.FixtureUtils (
+  AnalysisTestFixture (AnalysisTestFixture),
+  FixtureArtifact (FixtureArtifact),
+  FixtureEnvironment (NixEnv),
+ )
+import Path (reldir)
 import Strategy.Cargo qualified as Cargo
-import Test.Hspec
-import Types (DiscoveredProjectType (..), GraphBreadth (Complete))
+import Test.Hspec (Spec)
+import Types (DiscoveredProjectType (..))
 
 rustEnv :: FixtureEnvironment
 rustEnv = NixEnv ["rustc", "cargo"]
@@ -37,9 +43,13 @@ fd =
       [reldir|rust/fd/|]
       [reldir|fd-8.3.0/|]
 
--- These numbers can change as the dependency tree for sharkdp changes.
--- The fix for now is to just update the numbers when the spec breaks
 spec :: Spec
 spec = do
-  testSuiteDepResultSummary bat CargoProjectType (DependencyResultsSummary 146 29 268 1 Complete)
-  testSuiteDepResultSummary fd CargoProjectType (DependencyResultsSummary 74 25 145 1 Complete)
+  testSuiteHasSomeDepResults bat CargoProjectType
+  testSuiteHasSomeDepResults fd CargoProjectType
+
+-- The dependency number can change as the dependency tree for following changes,
+-- This is ramification of how `rust` builds are analyzed.
+--
+-- testSuiteDepResultSummary bat CargoProjectType (DependencyResultsSummary 146 29 268 1 Complete)
+-- testSuiteDepResultSummary fd CargoProjectType (DependencyResultsSummary 74 25 145 1 Complete)