+++ test/App/Fossa/Config/Utils.hs
@@ -38,6 +38,7 @@ itShouldLoadFromTheConfiguredBaseDir parser loadConfig =
                 , configTargets = Nothing
                 , configPaths = Nothing
                 , configExperimental = Nothing
+                , configVendoredDependencies = Nothing
                 , configTelemetry = Nothing
                 }
         maybeConfigFile <- loadConfig options